requirements.txt
feedparser==6.0.10
python-telegram-bot==20.7
python-dotenv==1.0.0
